refactor address utils
